WIP: Initial working proof-of-concept with updated docs.
